Update RadixPartitionOperator to use uint32_t for mask and include HashKernels
Changed the type of `mask_` from `uint64_t` to `uint32_t` in `RadixPartitionOperator.h` for better alignment with partitioning logic. Additionally, replaced the inclusion of `TargetSpecific.h` with `HashKernels.h` in `RadixPartitionOperator.cpp` to streamline dependencies and improve code organization.

+++ src/Common/RadixShuffle/RadixPartitionOperator.h
@@ -77,7 +77,7 @@ private:
     int K_;
     bool use_swwc_;
     int batch_;
-    uint64_t mask_; ///< P − 1 (P must be a power of two).
+    uint32_t mask_; ///< P − 1 (P must be a power of two).
     size_t max_cap_;
 
     std::vector<IScatterColumn *> cols_;
